Add encoding and decoding functionality to 'endigit' package
This commit implements encoding and decoding functionality in the `endigit` package. The Diff adds a new Go module with several required dependencies. The `endigit.go` file defines the `Config` and `Digit` structs for digit encoding, along with `NewDigit` function to initialize a new `Digit` instance. The `Encode` and `Decode` methods implements the encoding and decoding of digits. A set of test cases are also added in `endigit_test.go` to verify the functionality.

Moreover, the `.gitignore` file was modified to ignore all files with a whitelist (allowlist) approach, allowing Go files,